# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Revolut-inspired reference fintech platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /onboarding/start:
    post:
      operationId: startOnboarding
      tags: [onboarding]
      summary: Start onboarding session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartOnboardingRequest"
      responses:
        "200":
          description: Onboarding session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingSession"
  /onboarding/session/{sessionId}:
    get:
      operationId: getOnboardingSession
      tags: [onboarding]
      summary: Get current onboarding session state
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Onboarding session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingSession"
  /onboarding/session/{sessionId}/step:
    post:
      operationId: submitOnboardingStep
      tags: [onboarding]
      summary: Submit an onboarding step answer
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingStepSubmission"
      responses:
        "200":
          description: Next step or completion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingSession"
  /onboarding/steps:
    get:
      operationId: getOnboardingSteps
      tags: [onboarding]
      summary: Get full onboarding step tree (for rendering)
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [personal, business]
      responses:
        "200":
          description: Onboarding step definitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OnboardingStep"

  # KYC
...
  /kyc/submit:
    post:
      operationId: submitKycDocument
      tags: [kyc]
      summary: Submit a KYC document for verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycDocumentSubmission"
      responses:
        "202":
          description: Document submitted for review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycStatus"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    StartOnboardingRequest:
      type: object
      properties:
        type:
          type: string
          enum: [personal, business]
        countryCode:
          type: string
      required: [type, countryCode]
    OnboardingSession:
      type: object
      properties:
        sessionId:
          type: string
        type:
          type: string
          enum: [personal, business]
        currentStepId:
          type: string
        completedSteps:
          type: array
          items:
            type: string
        answers:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [in_progress, completed, abandoned]
        nextStep:
          $ref: "#/components/schemas/OnboardingStep"
      required: [sessionId, type, currentStepId, completedSteps, answers, status]
    OnboardingStepSubmission:
      type: object
      properties:
        stepId:
          type: string
        answer:
          additionalProperties: true
      required: [stepId, answer]
    OnboardingStep:
      type: object
      properties:
        stepId:
          type: string
        screenName:
          type: string
        questionText:
          type: string
        questionType:
          type: string
          enum: [select, multiselect, text, date, phone, upload, otp, country_picker, info]
        options:
          type: array
          items:
            $ref: "#/components/schemas/StepOption"
        validationRules:
          type: array
          items:
            type: string
        requiredDocuments:
          type: array
          items:
            type: string
        countryConstraints:
          type: array
          items:
            type: string
        branchingLogic:
          type: array
          items:
            $ref: "#/components/schemas/BranchRule"
        isRequired:
          type: boolean
        helpText:
          type: string
      required: [stepId, screenName, questionText, questionType, isRequired]
    StepOption:
      type: object
      properties:
        value:
          type: string
        label:
          type: string
        icon:
          type: string
      required: [value, label]
    BranchRule:
      type: object
      properties:
        condition:
          type: string
        nextStepId:
          type: string
      required: [condition, nextStepId]
    KycStatus:
      type: object
      properties:
        userId:
          type: string
        overallStatus:
          type: string
          enum: [not_started, pending, in_review, approved, rejected]
        identityStatus:
          type: string
          enum: [not_started, pending, in_review, approved, rejected]
        addressStatus:
          type: string
          enum: [not_started, pending, in_review, approved, rejected]
        livenessStatus:
          type: string
          enum: [not_started, pending, in_review, approved, rejected]
        submittedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        rejectionReason:
          type: string
      required: [userId, overallStatus, identityStatus, addressStatus, livenessStatus]
    KycDocumentSubmission:
      type: object
      properties:
        documentType:
          type: string
          enum: [passport, national_id, drivers_license, utility_bill, bank_statement, company_registration]
        documentFront:
          type: string
          description: Base64 or URL of front image
        documentBack:
          type: string
          description: Base64 or URL of back image (optional)
        selfieImage:
          type: string
          description: Base64 or URL of selfie
      required: [documentType, documentFront]
...
    Notification:
      type: object
      properties:
        notificationId:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [payment_received, payment_sent, card_transaction, kyc_update, system, security_alert, promotion]
        title:
          type: string
        body:
          type: string
        isRead:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
      required: [notificationId, userId, type, title, body, isRead, createdAt]
...