# lib/api-zod/src/generated/types/onboardingSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Revolut-inspired reference fintech platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OnboardingSessionAnswers } from "./onboardingSessionAnswers";
import type { OnboardingSessionStatus } from "./onboardingSessionStatus";
import type { OnboardingSessionType } from "./onboardingSessionType";
import type { OnboardingStep } from "./onboardingStep";

export interface OnboardingSession {
  sessionId: string;
  type: OnboardingSessionType;
  currentStepId: string;
  completedSteps: string[];
  answers: OnboardingSessionAnswers;
  status: OnboardingSessionStatus;
  nextStep?: OnboardingStep;
}